pyproject.toml
[tool.poetry]
name = "image-classification-with-lightning"
version = "0.1.0"
description = ""
authors = ["OTRAY271 <mojayarto@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
torch = {version = "^2.4.1+cu124", source = "torch_cu124"}
torchvision = {version = "^0.19.1+cu124", source = "torch_cu124"}
torchaudio = {version = "^2.4.1+cu124", source = "torch_cu124"}
lightning = "^2.4.0"
hydra-core = "^1.3.2"
scikit-learn = "^1.5.2"
wandb = "^0.18.2"


[[tool.poetry.source]]
name = "torch_cu124"
url = "https://download.pytorch.org/whl/cu124"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
